+++ src/classes/utils/HTMLMarkdown.ts
@@ -1,0 +1,9 @@
+import { MarkdownString } from "vscode";
+
+export default class HTMLMarkdown extends MarkdownString {
+	constructor(value?: string | undefined, supportThemeIcons?: boolean | undefined) {
+		super(value, supportThemeIcons);
+
+		this.supportHtml = true;
+	}
+}